+++ FusionIIIT/applications/hr2/models.py
@@ -614,8 +614,8 @@ class LeaveApplicationNew(models.Model):
     )
     nominee_responded_at = models.DateTimeField(null=True, blank=True)
 
-    medical_certificate = models.CharField(max_length=200, blank=True)
-    attachment_file = models.CharField(max_length=200, blank=True)
+    medical_certificate = models.FileField(upload_to='hr/leave/', blank=True, null=True)
+    attachment_file = models.FileField(upload_to='hr/leave/', blank=True, null=True)
 
     applied_date = models.DateField(auto_now_add=True)
     leave_balance_before = models.DecimalField(max_digits=5, decimal_places=1, null=True, blank=True)
@@ -963,6 +963,23 @@ class AppraisalFormNew(models.Model):
     rating = models.CharField(max_length=20, blank=True)
     status = models.CharField(max_length=20, choices=STATUS_CHOICES, default='PENDING')
     remarks = models.TextField(blank=True)
+
+    assigned_reviewer_role = models.CharField(max_length=20, blank=True)
+    assigned_reviewer = models.ForeignKey(
+        ExtraInfo,
+        on_delete=models.SET_NULL,
+        null=True,
+        blank=True,
+        related_name='assigned_appraisals_new',
+    )
+    assigned_by = models.ForeignKey(
+        ExtraInfo,
+        on_delete=models.SET_NULL,
+        null=True,
+        blank=True,
+        related_name='appraisal_assignments_made',
+    )
+    assigned_at = models.DateTimeField(null=True, blank=True)
 
     submitted_at = models.DateTimeField(auto_now_add=True)
